icontrol driver sync to reference service loadbalancer
Issues:
Fixes #81

Problem:
The agent_manager is refreshing the state of a loadbalancing service
because it has found a loadbalancing object in the pending state.
The sync() method in icontrol_driver refers to service['pool']
which refers to the old service model. This must be changed to
service['loadbalancer'].

Analysis:
Change service['pool'] to service['loadbalancer']

Tests:
Setup/teardown service.