Add conf.f5_common_external_networks to agent configuration report
Issues:
Fixes #85

Problem:
The LBaaSv2 needs to make a determination as to whether or not a particular
network tenant-id is valid for the service so that it can fail close
to the request to buid a service.  This flag is necessary to determine if
a network with the "router:external" flag is set to 'True' will be created
in '/Common'

Analysis:
Add self.conf.f5_common_external_networks to agent configuration report.

Test:
Manual