Add exception handling around prep_service_networking()
Issues:
Fixes #84

Problem:
There are many opportunities for the network construction to fail when
creating a loadbalancing service. This provides a mechanism to report
an error state back to the driver.

Analysis:
Add a try-catch around prep_service_networking() and report a loadbalancer
in ERROR state to the driver.

Test:
Manual