F5 agent running prevents creation of HA Proxy loadbalancers
Issues:
Fixes #88

Problem:
When the haproxy lbaas service is invoked the first thing that
will happen is to find an agent to perform the action. It queries the
database for all active agents of type AGENT_TYPE_LOADBALANCERV2.
The result will be all active loadbalancers, including the f5 loadbalancer.
In the code printed above, it then queries every active agent to see if
the device driver reported in the config matches the one passed in to the
method. The F5 LBaaS driver does not define this key in the agent config
and so the exception results when the F5 LBaaS agent is processed.

This happens when both the LBaaSv2 f5 driver is running in conjunction
with another LBaaSv2 driver that uses the ChanceScheduler.

Analysis:
Add the device_driver attribute to the agent_configuration.

Tests:
Ran manual tests.