Default pool not set on virtual server when creating pool without name parameter
Issues:
Fixes #104

Problem: The agent sets the default pool attribute of a BIG-IP virtual
server when an LBaaS pool is created. If the lbaas command used to create
does not include the optional --name parameter, the virtual server pool
attribute is set to an empty string.

Analysis: Modified the agent to define a valid value for the pool name. If
the name parameter is included in the command, that name is used. If the
name parameter is not used, the agent constructs the name using the tenant
ID and prefix.

Tests: Command scripts.