Refactor/cleanup tenant folder create/delete
Issues:
Fixes #132

Problem:
Need to make sure that exceptions that occur during folder creation
are handled appropriately.

Analysis:
Make sure that exceptions for folder and route domain creation and deletion
are captured and handled. On deletion, attempt to purge folder contents.
On creation, make sure that the loadbalancer is set to error state.

Extend the definition of resource helper to include collections of resources.

Extend the system helper to add functionality to purge folder contents and to
log warnings if purging orphaned folders.

Tests: