Virtual server creation only does Auto SNAT
Issues:
Fixes: 145

Problem:
Although we are creating a SNAT pool by default as specified by our configuration, virtual
servers are created with snat mode auto snat. If a SNAT pool gets created, then the SNAT
pool addresses should be used for SNATing.


Analysis:
Modify the service adapter to create a VIP that includes the SNAT pool
name if a SNAT pool is available.

Test:
Traffic and functional