SNAT creation failure for 'Common' network
Issues:
Fixes #156

Problem:
When the VIP network is assigned in the common folder the creation of a SNAT address
is successful; however, the SNAT pool creation in the tenant folder fails b/c the
code attempts to assign a translation from the tenant folder, and not the one created
in the Common folder. Since there is no SNAT translation address in the tenant
folder the pool creation fails.

Analysis:
Create the pool by correctly passing in the network folder which contains the
SNAT translation address.

Tests:
Set the VIP network to shared.
Run Traffic tests.