Unable to create more than one SNAT in a pool
Issues:
Fixes #161

Problem:
The snat code creates a translation address and a snat pool to use for
the virtual server. However, if the snat pool already exists the previously
added member is overwritten. So if a user specifies more that one SNAT
address, only the last one created is in the list.

Analysis:
The bug expected a pool creation error to occur in the event that
the pool already existed and the error code is 409.  The resource
manager code interprets a conflict as an update and proceeds.
This fix explicitly checks for the snatpool.  If it exists,
it is loaded and the members are updated with the most resently
created snat translation element.

Tests:
Traffic 2-armed/ and Common network.