Info-dump tool needed as part of the agent
Issues:
Fixes #241

Problem:
This is a new tool in the agent to bundle log and configuration
information together for easy access in troubleshooting an issue.

Analysis:
The script is install to /usr/bin/f5 when the f5-openstack-agent is
installed. I've updated the README to instruct a user how to run this
tool and some of the caveats about using. Namely, the logs and config
files may have sensitive information and should not be posted to a
public forum out-of-hand, without first scrubbery.

Tests:
Create a set of functional tests to get 100% coverage on this tool. All
tests pass. Note that these tests do no run in a virtualenv, as the
f5-os-testrunner tools (used in-house) creates a virtualenv to run these
tests.