Deleting a loadbalancer removes all elements in a tenant partition
Issues:
Fixes #228

Problem:
Create two loadbalancers in the same tenant. Delete one of the loadbalancers.
The tenant folder and remaining loadbalancer are removed from the BIG-IP.

Analysis:
The tenant cleanup call expects to be called when the folder is ready to
be removed; however, it is called whenever a loadbalancer is removed.  If there
are remaining objects in the folder, they will be destroyed too.  The
tenant cleanup should not remove everything in the folder.

Tests: