Agent should run as neutron user instead of root
Issues:
Fixes #227

Problem:
Service definition for f5-openstack-agent currently run as root. It is correct,
and requested by a customer, to run as user 'neutron'.

Analysis:
Change the service definition to uncomment the user neutron.  Change
the pre-install script to modify the log file ownership if it already
exists.

Tests:
Ran installation tests.