Route domain id allocation failure.
Issues:
Fixes #299

Problem:
When creating a new route domain the agent gets the set of all route
domains on the BigIP, and selects the lowest available route domain id
available.

This allocation fails when there is a hole in the list of existing route
domains b/c the lowest available route domain id is not returned.

Analysis:
Return route domain id at the end of the function.

Tests:
Create a loadbalancer/listener on BIGIP system where a hole in the route domain
id list exists.
Reran Tempest TLS test suite.