Issue #302 - StatsHelper should cast the stats as integers and deal with 'nan'
Issues:
Fixes #302

Problem:
The stats that are returned by the get_global_stats() are strings
and are blindly cast as integers through out the code base.
The problem is that in some circumstances these can be 'nan' and
the int() call will throw.

We should cast these to ints in the get stats function so people
don't need to worry about their types later.

Analysis:
* Cast the stat to int and catch any ValueError exception
* If we get the exception we set the value to 0

Tests:
* Wrote new unit tests