traffic_group distribution not working for scalen deployments
Issues:
Fixes #321

Problem:
When the agent starts it collects all provisioned traffic groups on the BIG-IP,
removing only the local_only traffic group. When a load_balancer is requested
the traffic_group is intended to be populated from a the collected traffic_group
list with tenant affinity. The traffic_group for the virtual_address is always
/Common/traffic-group-1.

Analysis:
Add the traffic group to the service object when it is discovered in
common_service_handler.

Tests: