tagMode should be specified when creating a vlan inteface
Issues:
Fixes #325

Problem:
VLAN tagMode needs to be specified when creating a VLAN interface
if that interface is to be tagged.

Analysis:
Try to always set the tagMode to 'service', if an exception is
caught with type TagModeDisallowedForTMOSVersion, then retry to
create the interface on the vlan without tagMode, as is the
case for 11.5.

Tests:
Unit testing, test/test_vlan