This commit searches for services that are retrieved from neutron and checks for the existence of the service.  If there is an issue:
folder does not exist
virtual address does not exist
service listeners do not exist
service pools do not exist

It then checks if the reason for not finding these is that an object
is using the openstack specified name for the name on the bigip.  It will
delete these objects on the BigIP.

A subsequent sync state is called that will redeploy the service with the
new names.