Add service adapter for l7policies and l7rules.
Issues:
Fixes #408

Problem:
In order to support L7 Content Switching in the agent, we must be able
to adapt the policies and rules in the service object to conform to the
bigip's view of these objects. This is the realm of the service adapter.

Analysis:
This adapter is a class, instead of a class method as before, that
inherits from ServiceModelAdapter and manages the translation. Each type
of object under a policy is also a class: Rule, Action, Condition. Those
objects are used as storage for the attributes needed to deploy those
objects on teh device.

Tests:
Created a set of functional tests, which use a real bigip. They validate
that the output of the adapter is as expected and that we are able to
deploy it on the bigip.