Adding environment prefix to rule actions, so we can recreate the pool id correctly, if needed.
Issues:
Fixes: #425

Problem:
The policy is not deploying properly when containing a rule
action that redirects to a particular pool.

Analysis:
Modified action class to have an environment prefix input parameter.
This is used to link the action redirect to a particular pool.

Tests:
Will run tempest tests against this code