L7Policy Adapter does not excise rule with a PENDING_DELETE status
Issues:
Fixes #430

Problem:
When adapting a policy and rules, the l7policy adapter should expect to
see a rule in a PENDING_DELETE operating status. This would mean to
ignore that rule when it does the translation for deployment on the
device.

Analysis:
Changed the instantiation of a condition to switch on the provisioning
status of the openstack rule. If the status is PENDING_DELETE, that
condition is not added to the translated l7 policy object.

Tests:
Wrote a new functional test to ensure this occurs as we expect.