Disconnected Service tests regression
Issues:
Fixes #434

Problem:
Disconnected service tests check for the existence of virtual address,
and virtual server based on the old naming scheme.  The object 'name'
in the service object is not used for naming, the UUID is.

The tests for disconnected network require that the virtual adddress
object creation be checked separately from listener creation.

Analysis:
Update the URIs to contain the UUIDs in the path and not the OpenStack
object name.

Create a new URI set for a virtual address.  This should be present
whenever a loadbalancer object is created, except in the cases
where the disconnected network feature is disabled and there is not
a segment id.

Tests:
disconnected network tests