Disconnected Service tests regression
Issues:
Fixes #434

Problem:
Disconnected service tests check for the existence of virtual address,
and virtual server based on the old naming scheme.  The object 'name'
in the service object is not used for naming, the UUID is.

Analysis:
Update the URIs to contain the UUIDs in the path and not the OpenStack
object name.

Tests:
disconnected_service