L7Policy adapter needs to have the logic to fully translate OpenStack policies and rules
Issues:
Fixes #452

Problem:
The l7policy adapter does not pay attention to policy provisioning
status nor does it do much by way of guessing what the work should be
for the BIG-IP. We need to make it a little more intelligent by having
it inspect provisioning status of the policies and rules and it will
inform the caller about the work that needs to be done. It will either
produce a JSON object for the caller to deploy onto the BIG-IP or it
will raise an exception, and if the exception is of a certain, the
caller will know that we can delete the wrapper policy on the device.
All other exceptions will be treated as errors.

Analysis:
Modified the policy adapter to take mostly full control of translation.
It is now aware of policies that are being deleted and rules that are
being deleted. It is smart enough to know when the wrapper policy should
be deleted completely.

Tests:
All tests that discovered this issue and address it are in the driver
repo and running as tempest tests.