Disconnected network tunnel remains after deleting a loadbalancer
Issues:
Fixes #293

Problem:
When creating and deleting a loadbalancer in global routed mode,
the tenant folder is not deleted because the disconnected network
is hanging around.

Analysis:
When the folder cleanup is invoked the delete_network() for
disconnected network is only called when there are route domains
present in the folder; otherwise, the deletion of disconnected
network is skipped.

Code inspection revealed that the create_network and delete_network calls
in disconnected_service are not symmetric.

create_network() invokes network_helper.create_tunnel whereas
delete_network() directly loads the tunnel object via bigip.tm
and does a .delete().

Tests:
    test_disconnected_service_creation.py::test_featureoff_withsegid_lb PASSED
    test_disconnected_service_creation.py::test_withsegid_lb PASSED
    test_disconnected_service_creation.py::test_featureoff_withsegid_listener PASSED
    test_disconnected_service_creation.py::test_featureoff_nosegid_lb PASSED
    test_disconnected_service_creation.py::test_featureoff_nosegid_listener PASSED
    test_disconnected_service_creation.py::test_withsegid_listener PASSED
    test_disconnected_service_creation.py::test_nosegid_lb PASSED
    test_disconnected_service_creation.py::test_nosegid_listener PASSED
    test_disconnected_service_creation.py::test_nosegid_listener_timeout PASSED
    test_disconnected_service_creation.py::test_nosegid_to_segid PASSED
    test_disconnected_service_creation.py::test_featureoff_grm_lb PASSED