Makefile targets need to use virtualenv for internal testing #470
Issues:
Fixes #470

Problem:
A test that is run via the Makefile for internal testing should use
a virtual environment so that it can install some of the internal
test packages and not conflict or update the base system. It should
also run the tools inside that virtual environment.

Analysis:
* Update the make file to create a virtualenv
* Use the virtualenv for the commands that need it
* Add a clean target to remove test artificats in tox and make

Tests:
* make functest