Deleting last VLAN in VCMP scenario results in all VLANs being removed from guest
Issues:
Fixes #447

Problem:
In a vCMP enabled scenario we are seeing an issue that when the last
load balancer in the system is deleted then all VLANs assigned to the
VCMP guest are removed, including those not managed but the Openstack
agent.

This leads to a situation where the management VLAN is removed and the
guests are left in a disconnected state.

Analysis:
Found that the VLANs were being wiped by a bad modify call in the vcmp
module. Changed the code to capture the existing vlan list, then remove
the element we want, and perform the modify call.

Tests:
All vcmp tests run pass