Issues: Fixes #418
Problem:
Starting the agent.py, which is the default service for the F5 openstack
agent, would result in an unhandled ImportError exception that would not
report to either the /var/log/ logs.

Analysis:
This logic will assure that the message is sent to the /var/logs and dump the
tracedump into the CLI when appropriate (debug mode is set in the
f5-openstack-agent.ini file).

Tests:
agent.py's test_agent.py under the respective /test:
./f5-openstack-agent/f5_openstack_agent/lbaasv2/drivers/bigip/test/test_agent.py