Issues: Fixes #418
Problem:
The init.d script is not handling the proper reporting of missing
dependencies when the agent is being initialized.  Here, the expected
behavior is that a log message should be logged stating that the agent
is not installed to an indeterminant (default) log (logger).

Analysis:
This should produce a log and dump into the CLI terminal debugging
information for any user.  A user may have to run thing several times to
get all dependencies, but they can at least get the list this way, if
they choose not to use the standard documentation.

Tests:
This is a shell script and is not all that conducive to testing...