Agent takes down virtual server on listener update.
Issues:
Fixes #565

Problem: Updates to listeners(neutron) cause the agent to destroy the virtual and
recreate it. This results in an outage and service disruption.

Analysis: The original code to destory/re-create was developed to handle updating
the name when the virtual server was named on the BIG-IP using the LBaaS name.
Because the agent code was changed to use the project name when naming objects,
the virtual server does not need to be destroyed and name updates can be
applied directly to the description.

Tests: test_listener_update.py