Issues: Fixes #418
Problem:
The issue of the init.d script not exiting as expected with the correct,
printed output into the CLI for start, stop, restart.  Have already
fixed the status in the past.

Analysis:
Am introducing additional logic to test the status of the daemon prior
to any start or stop.  If the daemon was already running, then it will
simply state that it was&is running without 'trouncing' the original
PID.
If it was already stopped when 'stop' is executed, then it will state
that it was already in a stopped state and state that it is stopped.  If
it was running when stop is issued, then it will simply state that it is
stopped or the particular error coded type, and state that it is still
running (if so).

Lastly, if the daemon is called from an unrecognized argument,
unprivilidged
user or the improper number of arguments are given, then it will exit
with the
proper exit statement and code.

Tests:
Have tested the following manually as this requires a full devstack or
similar build on ubuntu:
- start; restart; stop
$ sudo service f5-oslbaasv2-agent start;sudo service f5-oslbaasv2-agent
restart;sudo service f5-oslbaasv2-agent stop
(0) Service is running
(0) Service is running
(7) Service, f5-oslbaasv2-agent, is not running
- from stop, status; start; status; #kill# status; stop; start; start
$ sudo service f5-oslbaasv2-agent status
(3) Service f5-oslbaasv2-agent is not running!
$ sudo service f5-oslbaasv2-agent start
(0) Service is running
$ sudo service f5-oslbaasv2-agent status
(0) Service f5-oslbaasv2-agent is in an OK Status!
$ kill -9 $(cat /var/run/neutron/f5-oslbaasv2-agent.pid)
$ sudo service f5-oslbaasv2-agent status
(1) Service f5-oslbaasv2-agent is dead and /var/run pid file exists!
$ sudo service f5-oslbaasv2-agent stop
Service is already in stopped status
(7) Service, f5-oslbaasv2-agent, is not running
$ sudo service f5-oslbaasv2-agent start;sudo service f5-oslbaasv2-agent
start
(0) Service is running
Service is Already Running: no action
(0) Service is running