Pool member patch fails with 'address may not be specified'
Issues:
Fixes #584

Problem:
It appears as though we are doing a 'modify' call to patch a member
and we are completely expanding the attributes of a previously
loaded member (with ** keyword argument unpacking). That operation
seems fit for an 'update' call, unless you only want to change a few
things about member, then it's a patch.  This issue does not affect
BIG-IP 12.1

Analysis:
Before calling modify pop the 'address' attribute.

Tests:
neutron_lbaas scenario tests.
Called member_update weight on 11.6.0 BIG-IP