Add device multiplex to nightly tests
Issues:
Fixes #594 #596 #597

Problem: We support a range of devices.  Our nightly infrascture
needs to automatically test each of these devices.

Analysis: I added the changes to the Makefile following the
architecture swormke used in the f5-openstack-lbaasv2-driver repo
which was itself following the practices developed by the velcro
team.   The entrypoint into the nightly build process is that
the buildbot worker container runs (the equivalent of):

cd f5-openstack-agent/systest && make functest

Tests:  The new code was tested by running the above in an
appropriate environement (testenv generated stack on the "lab"
openstack cloud).  This correctly caused 4 separate runs of the
same tests.  Undercloud tests were not relevant for these tests
so the 4 undercloud RULEs simply echoed their names.