base the $(PROJNAME) variable on remote origin.
Issues:
Fixes #603

Problem: The systest Makefile $(PROJECT) variable assumed
that FOO/systest/Makefile mapped "FOO" to the name of the
project being tested.   Buildbot runs a `git clone SPAM .`
command in a directory named "build".  This means that the
"FOO/systest/Makefile" pattern _actually_ maps to
"build/systest/Makefile".  Clearly the names of "PROJECT"
variables cannot be derived from this filesystem path.

Analysis:  The solution is to use the remote that the
project was cloned from to derive PROJNAME.

Tests:  This was tested manually by verifying that "PROJNAME",
when derived from `git remote show -n origin` does in fact map
to 'f5-openstack-agent' in the case of this repository.