Opflex member can get stuck in PENDING if operations time-out
Issues:
Fixes #589

Problem:
If a port member creation fails after a timeout the loadbalancer
gets put in ‘ERROR’ state, but the member is left in ‘PENDING_CREATE’.
The loadbalancer should actually be left alone and the member set to
‘ERROR’ provisioning status.

Analysis:
Fix the agent_manager handling of pending services to update the
status of the service in the Neutron.  Properly handle exceptions
of objects that are deleted when exceptions occur in the network
setup.

Tests:
test/functional/neutronless/loadbalancer/
test/functional/neutronless/disconnected/
test/functional/neutronless/member/

Manual testing consisted of:
Create loadbalancer service with all objects: listener, pool, members
Create member with and without existing port.
Create member on opflex, add vlan child segment manually and verify
that the member is connected.
Create loadbalancer on opflex network, add vlan child segment
manuall and verify that the virtual get's created.
Create loadbalancer on opflex network, don't add segment and check
that the loadbalacner goes to 'ERROR' state and that it can be
deleted.
Create a member on the opflex network, don't add segment and check
that the member goes to 'ERROR' state after timeout period and verify
it can be deleted.