label stack names with a timestamp to handle edgecases
Issues:
Fixes #620

Problem:  If stack teardown fails, AND the build is
rerun against the same code and device, then the stack
names can collide.

Analysis: By adding timestamps to the names we make this a nonissue,
provided one doesn't have the extremely unlikely case of all of the
above being true AND rerunning within the same second.  Given the
amount of time it takes testenv to run this is a practical impossibility
anyway.

Tests: This was manually tested in the local buildbot sandbox.