place unstable tests in the noregression manifest
Issues:
Fixes #611

Problem: The build infrastructure is not yet stable.  To
stabilize it we need to discover the root causes of the
instability.  It's possible that poorly managed (torn down,
or setup) tests introduce instability into the regression
run.   It's likely that tests that have this effect will
fail.

Analysis: To isolate causes of instability in the regression
infrastructure we are marking unstable tests "no regression"
in manifest files that are external to the test code itself.

Tests: The determination of which tests are unstable and stable
was conducted using the bbot sandbox:
https://bldr-git.int.lineratesystems.com/kevin/bbot-sandbox