exclude tests from regression by GB project
Issues:
Fixes #625

Problem:  The regression tests are unstable, and this instability
varies depending on the device tested against, and the version of
the code under test.

Analysis: In order to get a baseline regression set I exclude
unstable tests as a function of device and code version.

Tests: These tests were shown to fail using the buildbot sandbox.