the coverage command is now in the Makefile
Issues:
Fixes #627

Problem: The directive to produce coverage results
was part of the tox.ini, but all other args are
specified in the Makefile.

Analysis: This commit adds the coverage commands to the
Makefile with all of the other py.test parameters.

Tests: This was tested manually with the buildbot sandbox.