update_loadbalancer_status getting called twice
Issues:
Fixes #643

Problem:
The update_loadbalancer_status call gets called twice after
deploying a service. This results in a database deadlock
occurring on the driver side, which was filed in the driver
as issue 472. The driver side fix is to handle database deadlocks
by catching the exception and retrying the operation.

Analysis:
Need to remove the extra call, that was included as part of a
bad merge.

Tests:
L7 tempest tests.