L7PolicyCreationException raised in creation of l7policy with rule
Issues:
Fixes #649

Problem:
When created a policy with a rule, in assure_l7rules_created, we
retrieve a particular parent policy for a rule, but the key we retrieve
it with is not correct. It is currently set as 'l7policy_id', but the
correct key is 'policy_id'. This is a reference to the id of the parent
policy from within the rule's service object.

Analysis:
Reference to wrong rule dict key 'l7policy_id' changed to 'policy_id' in
get_l7policy_for_rule in lbaas_builder.py.

Tests:
Ran failing test against BIG-IP 12.1 undercloud and it passed as
expected.