F5 sdk dependency version should be pinned in agent setup.py
Issues:
Fixes #657

Problem:
We should restrict the version for this package to be able to better
test, control, and document the package that users are installing.
Another good thing about pinning against a specific sdk version is that
we are not subject to any failings that occur in the sdk project.

Analysis:
The f5-sdk version is not restrict or pinned in the setup.py to 2.3.1.
We'll move this as we see fit going forward when we need appropriate
releases of the sdk. This also coincides with a doc change to ensure the
proper released versions of the sdk and icontrol rest are referenced.

Tests:
pip installed the agent from my fork.