Ratio load balancing method not set for pools with weighted members
Issues:
Issue #659

Problem:
Creating pool members with weights does not change the pool's lb-method
from round robin to ratio. The lb-method for pools with weighted members
should be changed to ratio. Weights (ratios) for members are correctly
set, but the pool lb-method is not changed. Note that OpenStack LBaaS
does not have a ratio lb-alogrithm, but the agent should ignore the
lb-algorithm setting when a member weight is specified.

Analysis:
Modified the service_adapter to set the pool's lb algorithm when it sees
a member specified with a weight. The lb method is changed to
'ratio-member' when the lb method is ROUND_ROBIN and a member has a
weight specified. The lb method is changed to
'ratio-least-connections-member' if the lb method is LEAST_CONNECTIONS
and a member has specified a weight. If SOURCE_IP is the lb method, no
change is made regardless of a member having a weight or not.

Tests:
Created a set of unit tests for the service_adapter changes, and some
neutronless tests for the functional testing of the lb method change on
a pool. The functional tests do not run currently in nightly, but they
do pass. I have created a JIRA story to track adding these and other
similar tests to the nightly tests.