References to bldr-git locations in systest should be updated
Issues:
Fixes #681

Problem:
This only affects the nightly builds. The hostname for the local gitlab
server in Boulder changed in late April, and now the references should
be updated to the new hostname. The new hostname is
gitlab.pdbld.f5net.com. The makefile, the install_test_infra.sh script,
and the neutron-lbaas.tox.ini files should be updated accordingly.

Analysis:
Updated the reference to the new hostname with slashes in the path
before the project name instead of a colon character.

Tests:
Ran tests on my local buildbot sandbox environment.