Mitaka nightly runs are failing with missing projects folder in systest
Issues:
Fixes: #688

Problem:
The nightly test runs for the agent in mitaka are failing because the
Makefile is expecting a project directory with 'mitaka' in the name, but
there are only liberty directories. This means no tests are executing at
all and the tox command is failing with a pytest-meta error.

Analysis:
Renamed the liberty directories to mitaka in systest/projects

Tests:
Ran the tests on my local buildbot system. There are still failures, but
the tests are running now.