Member updates failing due to missing address
Issues:
Fixes #680

Problem: Updating a member causes unhandled exception.

Analysis: Agent method to update pool member must remove the IP
address of the member before updating the BIG-IP. The address
will not necessarily exist, and this will cause a KeyError exception.
Added optional parameter None so that pop() will not fail if the
address key is not in the dict.

Tests: