Activate neutronless tests except vcmp
Issues:
Fixes #699

Problem: Many of our tests are not yet incorporated into the automated
build infrastructure.  This change adds 4 types of "neutronless" tests
'member', 'pool', 'loadbalancer', and 'listener'

Analysis:  We delayed incorporation of these tests into the nightly
infrastructure because we are seeking to establish sources of test error.
We are incorporating these tests now, because we are (for the moment)
focusing on the driver tests to establish stability.

Tests:  I ran these tests in a builbot sandbox environment.