yHTTP profile incorrectly added to TCP listener with SOURCE_IP persistence
Issues:
Fixes: #660

Problem:
When adding a pool with SOURCE_IP persistence, if the associated
listener has TCP protocol, the fastL4 profile is removed and http and
oneconnect profiles are added. These profiles are not required for
source_addr persistence profiles and should not be added to the virtual
server. Note, http profiles are required for HTTP and APP_COOKIE
persistence, and are also added for listeners with HTTP, HTTPS, or
TERMINATED_HTTPS protocol types.

Analysis:
Added if then logic in listener_service.py to set http and oneconnect
profile only when listener protocol type is not TCP.

Tests:
Added a functional test in test_session_persistence.py test module to
ensure the correct profiles are added when a pool is attached to a
listener with session persistence defined.