Upgrade to testenv-all
Issues:
Fixes #719

Problem: The testenv tool and dependencies had to be managed individually,
it's now the case that the testenv-all wrapper package manages versions
(and formerly in systest-common) and templates.

Analysis: This is an upgrade to tooling, see:

https://gitlab.pdbld.f5net.com/bdo/testenv-all/blob/master/README.rst

for details.

Tests: I manually tested these changes in my local environment.