Makefile needs to be prepared for newton running in nightly
Issues:
Fixes #747

Problem:
The systest Makefile needs to be able to handle a branch name of
stable/newton. That means we need to sanitize the name to strip out the
forward slash. This is necessary because forward slashes can and will
introduce all sorts of fun issues when that branch name is used in say,
a directory path. Also, jenkins will not allow a job named with a
forward slash. So, we will sanitize that name, then to get the OpenStack
distro, we can strip the 'stable' string out of that sanitized branch
name. This allows us to match the branch name as closely as possible for
internal test reporting.

Analysis:
Created a SANITIZED_BRANCH variable in the makefile. This represents the
branch name without any forward slashes. That variable is used in path
names wherever necessary.

Tests:
Ran tests on my jenkins container.